+++ frontend/src/views/App.svelte
@@ -3,6 +3,7 @@
   import Header from "./Header.svelte";
   import Content from "./Content.svelte";
   import Errors from "./Errors.svelte";
+  import Updater from "./Updater.svelte";
 
   const modalSettings = {
     key: 'modal',
@@ -97,5 +98,6 @@
     <Header />
     <Content />
   </ModalContext>
+  <Updater />
 </main>