+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Notification when a new version of Wombat is available to download.
+
 ### Fixed
-- support mac case-sensitive filesystem. Thanks to [@Azuka](https://github.com/Azuka)
+- Support mac case-sensitive filesystem. Thanks to [@Azuka](https://github.com/Azuka)
 
 ## [v0.2.2] - 2020-11-12